Consolidate classpath. Added more javadocs
